# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "API documentation description"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/attributes": {
      "get": {
        "description": "Retrieve the full list of attributes",
        "tags": [
          "Attributes"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": [
              "Attribute"
            ],
            "schema": {
              "$ref": "#/definitions/Attribute"
            }
          }
        }
      }
    },
    "/concentrators": {
      "get": {
        "description": "Retrieve the full list of concentrators",
        "tags": [
          "concentrators"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array",
            "schema": {
              "$ref": "#/definitions/Concentrator"
            }
          }
        }
      }
    },
    "/issues": {
      "get": {
        "description": "Retrieve the full list issues",
        "tags": [
          "Issues"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": [
              "Issue"
            ],
            "schema": {
              "$ref": "#/definitions/Issue"
            }
          }
        }
      }
    },
    "/issues/groups": {
      "get": {
        "description": "Retrieve the groups of issues",
        "tags": [
          "Issues"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": [
              "Issue"
            ],
            "schema": {
              "$ref": "#/definitions/Issue"
            }
          }
        }
      }
    },
    "/issues/{id}": {
      "get": {
        "description": "Get the results of the study with the ID provided",
        "tags": [
          "Issues"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the study",
            "required": true,
            "type": "Integer"
          }
        ],
        "responses": {
          "200": {
            "description": "loop:name:result"
          }
        }
      }
    },
    "/packages": {
      "get": {
        "description": "Retrieve the full list of packages",
        "tags": [
          "packages"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array",
            "schema": {
              "$ref": "#/definitions/Package"
            }
          }
        }
      }
    },
    "/performances": {
      "get": {
        "description": "Retrieve the full list performances",
        "tags": [
          "Performances"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": [
              "Performance"
            ],
            "schema": {
              "$ref": "#/definitions/Performance"
            }
          }
        }
      }
    },
    "/performances/{id}": {
      "get": {
        "description": "Get the performances results of the study with the ID provided",
        "tags": [
          "Performances"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the study",
            "required": true,
            "type": "Integer"
          }
        ],
        "responses": {
          "200": {
            "description": "loop:name:result"
          }
        }
      }
    },
    "/studies": {
      "get": {
        "description": "Retrieve the full list of studies",
        "tags": [
          "Studies"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": [
              "Study"
            ],
            "schema": {
              "$ref": "#/definitions/Study"
            }
          }
        }
      },
      "post": {
        "description": "Create study. (too complex too test it through API)",
        "tags": [
          "Studies"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "description": "ID of the study to return",
            "required": true,
            "type": "Integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Study",
            "schema": {
              "$ref": "#/definitions/Study"
            }
          }
        }
      }
    },
    "/studies/{id}": {
      "get": {
        "description": "Find a study by ID",
        "tags": [
          "Studies"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the study to return",
            "required": true,
            "type": "Integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Study",
            "schema": {
              "$ref": "#/definitions/Study"
            }
          }
        }
      }
    },
    "/users/authenticate": {
      "post": {
        "description": "Retreive user info  bearer token",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "example": {
                  "username": "xaxi",
                  "password": "password"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "description": "Retrieve the full list of users",
        "tags": [
          "users"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Array",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Attribute": {
      "type": null,
      "required": [
        "id",
        "name"
      ],
      "properties": {
        "id": {
          "type": "number"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Concentrator": {
      "type": "object",
      "required": [
        "id_concentrador",
        "lvcid"
      ],
      "properties": {
        "id_paquete": {
          "type": "number"
        },
        "lvcid": {
          "type": "string"
        }
      }
    },
    "Issue": {
      "type": null,
      "required": [
        "id",
        "name",
        "description",
        "group",
        "priority",
        "detect",
        "fix"
      ],
      "properties": {
        "id": {
          "type": "number"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "group": {
          "type": "number"
        },
        "priority": {
          "type": "number"
        },
        "detect": {
          "type": "array"
        },
        "fix": {
          "type": "array"
        }
      }
    },
    "Package": {
      "type": "object",
      "required": [
        "id_paquete",
        "nombre",
        "descripcion",
        "fecha_insercion"
      ],
      "properties": {
        "id_paquete": {
          "type": "number"
        },
        "nombre": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "fecha_insercion": {
          "type": "date"
        }
      }
    },
    "Performance": {
      "type": null,
      "required": [
        "id",
        "name",
        "description",
        "group",
        "priority"
      ],
      "properties": {
        "id": {
          "type": "number"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "group": {
          "type": "number"
        },
        "priority": {
          "type": "number"
        }
      }
    },
    "Study": {
      "type": null,
      "required": [
        "id",
        "name",
        "initialDate",
        "loops"
      ],
      "properties": {
        "id": {
          "type": "number"
        },
        "name": {
          "type": "string"
        },
        "initialDate": {
          "type": "date"
        },
        "loops": {
          "type": "number"
        },
        "currentLoop": {
          "type": "number"
        },
        "start": {
          "type": "number"
        },
        "concentratorsAmount": {
          "type": "number"
        },
        "tasksAmount": {
          "type": "number"
        },
        "taskFinishedAmount": {
          "type": "number"
        },
        "taskProgressAmount": {
          "type": "number"
        },
        "taskPedingAmount": {
          "type": "number"
        },
        "taskQueuedAmount": {
          "type": "number"
        }
      }
    },
    "User": {
      "properties": {
        "id": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "integer"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  },
  "tags": []
}